contracts: Make claimId a composite id
The system needs to ensure that every request can be claimed multiple
times. However, it must only be able to be claimed once by any account.

This was not true before, when the claim id was a simple counter. This
PR changes the id to be a hashed composite value, consisting of the
request id and the claimer. This allows for easy checks whether or not
the claiming account has already done so.