raisync: make event fetching more robust and adaptive
We start by fetching up to 1000 blocks at a time. If that
seems to be too much for the ETH node to handle, we reduce the number of
blocks and try again. Similarly, if the ETH node can handle it, we
increase the number of blocks to fetch in one call.

Inspiration and code taken from
https://github.com/raiden-network/raiden-services/blob/cce18ef4dc8cb0cbc5d1b75bf1d1fd97a74a5d64/src/raiden_libs/blockchain.py#L294