contracts: Fix request fund payout when challenger wins
Imagine a claim happens by Bob who did NOT fill the request. Now Charlie sees
that and challenges the claim. Non of them should get the request money on the
source chain, as it is currently done.

So the change is done so that only the claimer receives the request funds when a
claim is withdrawn. If Chalie would have been the filler, he would need to
create a seperate claim, which wouldn't be a problem.