frontend: Refactor web3 functionality to a service
The service is called MetaMaskProvider because MetaMask is the de facto standard of browser injected wallets.
But this also limits the use to MetaMask. This way we can use functionalities that is only supported by MetaMask, like switching the chain.
If we want to support other providers we should implement other classes which implement the new interface.