frontend: Refactor components to use vue 3 composition API
As we want to have a frontend that can be used in production, this sets up a proper architecture for it.
This restructures all components to the new composition API. So vue-class-components was removed.
Some functionalities are moved to composables, some to services. This makes everything more modular, loosely coupled and easier testable.
The EthereumProvider is now provided by the new Provide/Inject API. This allows to make it easily available to components that need it.