frontend: Use JSON file for app configuration
Having the configuration options all as environment variables as before,
can make the env files pretty big if need to add deployment info for several
chains. Another benefit of a json file is that it can be changed without the
need to rebuild the app. The json file could later get generated
automatically during contract deployment process.
The new json file is loaded on app initialization. I used a composable for
the fetching to have it decoupled.