Fix transaction end to end flow
It wasn't possible to achieve a full end to end transaction.
Configuration was outdated, so it has been updated. There were some
problems with transaction because of missed gas limit value, we use now
estimateGas function to add this value, also validityPeriod was missing,
for now it's hardcoded until we add a component in UI to add it to the
request. Also, this should also fix the issue of listening to network
change event, now a proper error message appears if an unsupported
network has been chosen. Other minor code linting or unneeded comments
have been fixed.
It worths mentioning that while a successful flow can happen, but we
still have two major problem with the wizard steps state and also with
network change in select boxes. These are covered in other issues.

Resolves:
Full functional transaction
https://github.com/beamer-bridge/beamer/issues/331