Map amount with token decimals
We are trying to reflect the real amount sent from frontend. So if the
user puts 1 TST to send, a really 1 TST is sent and not mapped to the
number of decimals. This allows also the use of decimals in numbers,
like a value of 3.243, it is laso converted into its proper value.

Resolves:
https://github.com/beamer-bridge/beamer/issues/261