Fix switching chain in request form
User was not able to use source chain properly when a switch happens.
Whenever a switch event occurs, and error is thrown because of security
requirements by metamask, and the user was then forced to make a manual
refresh to be able to use the switched newtork without errors.
Mainly, we use location.reload to go over different issues we have with
switching networks, and for the case of a successful switch, we keep a
copy of chainId in localStorage before reloading the page, and then
rehydrate the source chain selectbox with saved chain option so the user
doesn't need to rechoose it again. We need as an extra step to keep the
whole form in localStorage so user doesn't lose entered information when
reloading the screen on switching source network.
This also fixes the issue of current network validation on submit, as on
switching network, we reload the screen in all cases, and this retains
the chain selected in metamask the same as the one in select box.
Worths mentioning that we are now adding rpcUrl as a part of chain
configuartion so it's easy to add network to user Metamask if we
detected that network is not available for user.

Resolves:
https://github.com/beamer-bridge/beamer/issues/263
https://github.com/beamer-bridge/beamer/issues/256