Stop hiding form for unsupported chain
When user chooses an unsupported chain from metamask, an error message
was appearing instaed of the form. This behavior doesn't allow the user
to know what are supported chains or what should do with the app.
Message appears now on top of the form instead. It isn't optimum, but
it will be there until a notification system is implemented for the app.

Resolves:
https://github.com/beamer-bridge/beamer/issues/360