Fix progress steps logic
In the processing screen, user was asked to switch to target network on
metamask to be able to listen to fulfillment event and see the full
result on processing screen. For now, user doesn't need to make this, as
we create anotehr instance of ethereum provider for the target network
that doesn't need metamask involvement in that. This has lead to removal
of switch network step from UI, progress state and the logic behind.
Some UI changes has been provided as well to processing screen.

Resolves:
https://github.com/beamer-bridge/beamer/issues/348
https://github.com/beamer-bridge/beamer/issues/277
https://github.com/beamer-bridge/beamer/issues/257
https://github.com/beamer-bridge/beamer/issues/292