Limit permission for token transactions
For every request, user was asked to make first a transaction for
giving permission to access token balance. This was happening because
everytime the persmission was given over the amount going to be sent.
We changed that to give the permission over the whole token balance
instead. By that, the permission transaction will happend once, unless
the user has made a request for a sum more than the initial balance

Resolves:
https://github.com/beamer-bridge/beamer/issues/353