Add button to add token to metamask
Because user doesn't have an information about the token going to be
used in testing, so it will be hard to see how the system is working,
a button to add the token to metamask is available for the user to
easily watch the token in metamask.
A small fix for the imprint link repalcing the dot separator with a
bullet.

Resolves:
https://github.com/beamer-bridge/beamer/issues/463