agent: Cache latest block data
Currently black data are queried from RPC in handlers when required.
This is ineffective, as most of these requests will get the same return
value.

Instead, we add a mapping from chain id to block data to the `Context`,
in which the latest known block for each chain is stored. It is updated
via the new `LatestBlockUpdated` event.