Fix always successful requests issue
It appears that I removed the request Id value from request object by
error. I thought the I use the receipt but found that I was wrong. On
the other hand, receipt value is not required to be in request object,
so it has been removed. A slight change to the RequestCreated event
finding method implementation to make it more clear.