replace ethereum provider provision with store
Same strategy as for the configuration. We replace all former
injection/provision data with proper stores. Additionally we finally started to
use some additional advantages of stores. Namely for this case getters which
allow us to write some more convenient and better readable code. Though we are
just scratching the surface here.

There is a special case here because we need to make the provider instance
explicitly a shallow reference in the store. This is necessary because deep
inside the ethers.js library there is an issue of using a deeply proxied object.
It is not very nice to have such code restrictions/adaptions in our core. Though
we were not able to find a better solution for now. Furthermore this is
basically the same solution as before. With more input, we might be able to
improve here in future.
Cause by this, the store getters which are based on the `provider` state
explicitly state their return type to ensure that the shallow reference gets
"resolved" and no user of the getter has to access the `.value` first, as this
would be an awkward exception to the rest of all store interactions.

It has been decided to make all utility functions which depend on an established
connection to the Ethereum provider to return `undefined` if that is not the
case. This simply avoids the complexity to provide special treatment in these
simple functions. It is the responsibility of the caller to decide what to do in
case of `undefined`.

With the tests for the new store, it has become necessary to write a first mock
implementation. Because this mock targets the `EthereumProvider` interface there
is a shared mock within the `tests` directory. There is no `__mock__` in the
`src` directory because first, there are multiple implementations of the
`EthereumProvider`. For some parts of the code it doesn't matter which
implementation is used and the tests should not make a decision which of them to
use. In case that one will become removed in future, many tests can break. This
relates to the second point, that these objects get instantiated directly in the
tests and are not indirectly used by imports of the component under test.