refactor composables according to design decisions
The usage of composables and stores was heavily under discussion. Due to lack of
experience with the new composition API of Vue, we had a hard time to figure out
which approach is the best for clean and good testable code. Finally we came up
with a design decision proposal that got accepted by the whole team. This commit
message does not aim to repeat the proposals content (please checkout the
related issues and pull requests of the project). But because this is such
a huge commit there is the need to explain why.

Additionally to the above mentioned code design changes, a few other changes
have been made too.
First of all has each composable been split into its own TypeScript module. Some
composables are quite complex. It is "hard" to decide which composables go
together and which to split. In general there is no advantage in having big
merged composable modules.
Second the RequestFormInputs component does no more receive the request fee as
a property. With the completely refactored composable for the request fees, it
is simply complicated to forward these information as property without missing
the new won features. After all the RequestFormInputs component is already very
complex and is involved in all input values anyway.

Working on the RequestDialog and RequestFormInputs component has shown that both
are way too large and too complex. It is hard to work with them and even the IDE
has a lot to process for them. Improving this should be target of future
commits.