refactor ProgressStep and RequestProcessing
The here made changes help to make the components more strict and
simplified. It follows the concept of separating components into dump
visual ones without much logic and data and smart components with data
and logic, but barely any custom visuals.

The ProgressStep component is a leaf child in the component tree and
thereby a purely visual. In result the component should not contain any
logic or make assumptions about how it gets used. For this specific
component this means it should not be aware about any request state
logic and ordering. It should not decide if a step was successful or
not. It should just provide the visual features to render these states
to for the user. All data and the logic of success, warning etc. will be
made by a parent component. These changes also ensure that the
component becomes much better to re-use, removes unnecessary
dependencies (i.e. to the RequestState enumeration) and improves
testability by reducing decisions to simple booleans (in theory the
previous tests where not complete).
Furthermore the slot has become a plain test property. This might look
like a step backwards and against the goal to make this a universal
visual component. But in fact this component has a clear visual task.
And that is not only the coloring and design of the steps state, but
also to ensure that multiple steps can be listed together. Allowing the
injection of an arbitrary DOM element can lead to inconsistent renders.
As long as we don't need this flexibility, we better don't have it. If
this component would be exposed as a library, it could look different.

The RequestProcessing component started to turn into a logical
component. It now determine which steps have been completed or failed.
The template for the steps has become more generic with a loop in
combination with a computed property.