frontend: extract request summary to own component
Displaying the metadata as a summary of a request is pretty trivial but
includes a bit of template to render all information. Thereby it makes
sense to extract this into its own dump visual component with simple
tests.

In result the `RequestProcessing` component could become much simpler by
being a purely logical component that renders the progress steps
dynamically bases on the current state. Tests for this component become
very simple and straight forward for a single feature.

Finally both components become put together in a tiny visual component
that uses also the `Card` component. Tests are trivial.

As a side effect of this, the request state got decoupled from the
request metadata. This nesting was not so nice and with the new
component separation, it makes much sense to also separate the state for
this.