frontend: add unit tests for Imprint component
This includes a small refactoring of the component too. So far we were
using classes by DaisyUI to handle the open and close part. This is very
tough to test or rather not possible. As this effect is so simple, it
got replaced with a custom mechanism. The whole appearance is the same
as before.