frontend: Refactor transfer code into single composable
Before we had two composables, one for the fulfillment and one for the request.
They were not adding much value, because their error variables were not used.
This change also removes the request construction code from the RequestDialog to make it cleaner.

According to #605, we want to have methods for running our actions attached to a Transfer object in the future (which is kind of the Request object at the moment).
This should then be created in the new `useTransfer` composable. Having one composable facilitates these changes.

Also, now the errors of a sent request and the fulfillment are handled properly, in one variable of the `useTransfer` variable.