Integrate WalletConnect to Dapp
To support non Metmask users, WalletConnect support has been provided
through this PR. This still lacks the proper UI and some minor issues
are still around. The core idea here is two things, first is to
implement `EthereumProvider` interface just like `MetamaskProvider` and
then adapt the different functions with the slight differnces between
Metamask and WalletConnect. The second thing, is proposing the store
persistor as it was manadatory for keeping the state of the connected
wallet (mainly for WalletConnect), so the user is kept connected even
after reloading the screen. `pinia-plugin-persist` has been used for
this goal, and in configuration, `localStorage` has been chosen as the
storage technology (API) for browser. A new store `settings` has been
added to keep the connection state and a unit test has been created
for it.
Testing should be added for the provider classes and the proper UI
should be applied before merging this, including the disconnect from
wallet feature. We have also and issue with showing the source chain
value in select box after changing because window reloads and the value
resets and is no longer available after that! A fix will be pushed
later for this.

Resolves:
https://github.com/beamer-bridge/beamer/issues/472