frontend: add types for strong encoding
In will become necessary to preserve data to the users local storage.
This means data must be stored and later reloaded again. This process
requires to encode the data and decode it again. Therefore it is helpful
to define type restrictions to ensure this process works as intended.

First of all it is necessary to define which set of types can be
actually stored. For example it is not possible to store functions.
Next, any custom type/class/... that will become preserved needs to
define an encode function. In turn it must accept the exact same
encoding output it produces as input to its constructor to decode the
data again.
The type restriction is put on the constructor, because TypeScript does
not allow to use define `static` fields in an interface. Thereby
a static `decode` function can not be defined. In result only the
constructor remains as an option.
Moreover it is not possible to have a single interface to define the
constructor as well as anything else. The constructor is not part of the
instance type of a class.
The resulting type harness isn't as optimal as it could be with better
language support, but it will help to write better code that reduces the
chance to introduce compatibility issues.

The yet unused `decode` function that is necessary to apply the decoding
restrictions is intended to be used by the persistence layer at a later
stage. It is just important to have the whole encoding restrictions in
place now so they can be already applied where a future persistence is
planned.
# On branch frontend/feature/multi-step-actions
# Changes to be committed:
#   (use "git restore --staged <file>..." to unstage)
# 	new file:   frontend/src/types/encoding.ts
#
# diff --git a/frontend/src/types/encoding.ts b/frontend/src/types/encoding.ts
# new file mode 100644
# index 0000000..3c838c3
# --- /dev/null
# +++ b/frontend/src/types/encoding.ts
# @@ -0,0 +1,21 @@
# +type EncodableBaseTypes = boolean | string | number | undefined;
# +
# +type EncodableArray = Array<EncodableBaseTypes>;
# +
# +interface EncodableObject {
# +  [key: string]: EncodableBaseTypes | EncodableArray | EncodableObject;
# +}
# +
# +export type EncodableData = EncodableBaseTypes | EncodableArray | EncodableObject;
# +
# +export interface Encodable<T extends EncodableData> {
# +  encode(): T;
# +}
# +
# +export interface Decodable<T extends EncodableData> {
# +  new (data: T): Encodable<T>;
# +}
# +
# +export function decode<T extends EncodableData>(decodable: Decodable<T>, data: T): Encodable<T> {
# +  return new decodable(data);
# +}