frontend: add new concept of a step
A step describes a phase/section of an action. An example for such an action
will be a transfer that requires multiple steps from making a request
transfer all they way to seeing the fill event. The individual steps
keep track of in which phase the action is. It adds a description and
matches a callable function that is associated with this step.

The whole concept of steps gets explained in much more detail in our
architecture decision documentation.