frontend: extend chain and token types
The chain information within the configuration file does not also
include the chain identifier too. This simplifies the access to this
information without having to always look back to the key of the
mapping. It might probably happen that the whole mapping will be
replaced with a plain list. But for the matter of keeping changes more
minimal, we just add more (duplicated) information here.

Having the new concept of transfers in mind, it is not nice that chain
objects are incredibly large with all tokens included. The list of
tokens is expected to grow heavily. As we will need to preserve chain
information with a transfer, this would become highly inefficient.
Taking into account that transfers will need to be deeply reactive, this
just makes it worse.
To solve this problem, the both have been separated for each other. Only
in the configuration the chain dataset gets extended with a list of
tokens. This might become a target of change in future too.

Last but not least do tokens now also include the information about
their decimals. As we configure each individual token anyway, it is very
simple to add this piece of information too. We have had defined the
token symbol too, even it could have been fetched from the blockchain.
But having to fetch this data over and over again asynchronously is
a pita. To keep things simple and effective, the decimals get now also
configured.