frontend: add transfer as first multi-step action
The new transfer as first multi step action finally makes use of the
whole new concepts that got introduced earlier. Having the whole code
separation in place, it is very easy and straight forward to test the
functionality of the transfer. Even for later backwards compatibility
tests.
This also makes use of the improved typing.

Right now, there is a lot of redundant code, as in the actual Vue
components, a transfer is still named a request and gets handled by
composables, the metadata object and the request state. This will be
target next and makes everything clean again. Though it would be simply
to chaotic to put this all in one commit.