frontend: make transfer progress component generic
Due to the concept of multi-step actions, it makes sense to have
a generic component that allows to display any action's progress.
This moves the (still legacy) logic for determine which state is
active to the transfer status component.

Please note that though the concepts of the multi-step actions and the
components to display the progress are close related to each other, but
still kept independent. So will types of the core business logic not be
re-used here to avoid unnecessary dependencies in small visual
components. These components rather explicitly define which information
they need. Logical components that integrate business logic and visual
parts will translate. In most cases this should be trivial. In best case
all visual components and their tests remain untouched in most cases.