frontend: adapt components to use transfer action
This is the final step to introduce the transfer action into the actual
Vue components. Instead of using the transfer composable, it executes
the multi-step sequence.
As affect of this, the transfer status component also receives
a transfer as property now. Having this, it is straight forward to add
a transfer history in future.