frontend: add new UInt256 type
The Ethereum virtual machine its unsigned integer (`uint256`) values are
much bigger than the JavaScript number type can handle. Therefore it is
necessary to have a special type to represent this data and allow
operations on it.

The `ethers` library comes with a type called `BigNumber` that does all
the job. Though it is not a good practice to introduce such 3rd party
types into the core type system of the business application. Furthermore
does `BigNumber` not fulfill the `Encodable` interface. This is
necessary to properly persist data like transfers. There is a function
to convert a `BigNumber` to a JSON notation and backwards. So it
kind-off satisfies the `Encodable` and `Decodable` interfaces. Though
not correctly.
After all there is no other chance than wrapping up this type in
a custom. This custom type can define an interface that fits the needs
exactly like the application needs. This can be extended any time
necessary. And most importantly, the hidden library is very limited and
easily exchangeable without other modules recognizing it. The only spot
where `BigNumber` will be necessary again is when using `ethers` to
interact with the contracts. Here the new custom `UInt256` type needs to
be converted back to a BigNumber. But this is again very limited at
a special spot that is tight to `etheres` anyway.