frontend: add new token amount type
This is a nice convinience type that helps to interact with with token balances
and amounts. Therefore it combines the usage of the new `UInt256` type and
a specific token. The hereby most important information by the token is the
number of decimals. This gets used to parse user inputs as well as format
amounts properly.

After all this types hides some complexity and makes it easy to work, without
having to remember what a specific `UInt256` is about, what token relates to it,
calling `ethers` utility functions everywhere and so on.

Furthermore this type also follows the encoding rules and can be persisted as it
necessary to be used with a transfer.