frontend: apply new types for transfers
This makes use of the `UInt256` and `TokenAmount` types in context of
a transfer. After all these types have been initially added to solve the
problem of having encodable `BigNumber`s. Without them, the issue so far
was that all transferred value where off by a factor of `10^18`
(depending on the token decimals). To finally address this issue and
resolve all of the `TODO`s in the code, this commit is here. Though this
commit does not make broader refactorings of applying the new types also
at different locations that can benefit from it. This is simply for the
matter of having the commit limited in size. Further improvements
planned.

Because of the chain of encoding data, all more complex types that
include class the must be encoded need to be a class too, to support
this encoding function. For that reason the `requestMetadata` had to
become a little class that supports this. Though the
`requestFillMetadata` remains a simple type. This is the same of other
encodable by default types like `Token` or `Chain`.
This is also the reason why a static new function got added to the
`Transfer` class too. This is the case for many `Encodable` classes to
simplify the usage as the constructor must follow certain type
restrictions. This leads to some ugly back and forth de-/encoding
between static initializer and constructor, but that is the prize we
pay.

This also makes the tiny change of declaring a type `TransactionHash`
this is the same as `EthereumAddress`. So far it is just a simple link
to `string` and only improves the readability and purpose slightly. It
is imaginable to put actual type restriction here to ensure correct
data usage/parsing.