frontend: move main card shape to home view
The core of the whole application is within a single Card component, in
fact a "box" centered on the screen. Below it is an action button that
can change. The content within the box does also change (e.g. form
inputs or transfer progress).
In preparation for future design adaptions, the Home component becomes
responsible to define this main Card where all content will take place.
Additionally it defines a teleport target where any component that will
be rendered within the box can display and action button (e.g. connect
wallet, transfer funds, ...).
Moreover it must now also display the wallet connection status as that
gets rendered on top of the main box.