frontend: new tabs component
The new Tabs component is a visual layout component that is necessary
for the new structure of the application. It allows to have two tab
headers that can be clicked to switch the content of the tab.

The tricky part is to ensure that the tab content does not get unloaded while
switching between the tabs. This must be done using Vue's `<KeepAlive>`
component. In turn this requires to define the content via Vue's special
`<component>` component. Slots might have been more straight forward and
flexible, but unfortunately do not work for this.
Furthermore, the `<component>` usage is quite cumbersome and doesn't allow for
the most "elegant" code in the component using it. More details in the comment
at the code itself.

The Card component was in need of a little extension to hide any overflow. The
problem was that when displaying a Tab component within a card, the background
of the inactive tab header will draw over the border radius of the Card.