frontend: put request dialog into tabs component
This makes usage of the new Tabs component and starts to change the UI
according to the design changes. Though the activity tab is still empty
for now.

This causes changes how a new transfer is made. In fact it does no more
require to reload the whole page. This was also necessary because the
reload was done via an event which can't be handled like this anymore.