frontend: make tabs component generic over count
The `Tabs` component was fixed to have a left and a right tab. According to the
current design, this is the only use-case. Anyway this is a visual component.
Visual components should try be somewhat generic, especially for such obvious
features. What makes it even more desirable in this specific case is
that is simplifies the component a lot. Thereby the code gets even clear
while supporting a greater variety of use-cases.