frontend: fix reactivity of form inputs
In the past the source chain was known when the request form gets
rendered. This has changed now and it is undefined in the beginning. To
solve this problem, the source chain and source token inputs are using
`v-model`s for two-way communication now. Both connect to computed
references that also define a setter now. In case of the source chain
this allows to update the input when the wallet its connected chain
changes. And in case of the token it allows to get rid of the manual
`@input` handling.