frontend: change terminology from metadata to info
Before this commit we had transfer properties called `requestMetadata` and
`requestFillMetadata`. The problem is that we have the concept of encodable
data. Here we have multiple classes which define related data type for that
purpose. The general naming convention is to use the class name plus `Data`.
This leads to the awkward name `RequestMetdataData` which is less than optimal.
Therefore it has been decided to switch from `metadata` to `information` to
avoid this collision.

Furthermore we decided to rename the also not so nice `RequestFillInformation`
(after applying the above) to `FufillmentInformation` only. Due the context it
is clear what fulfillment we are talking about. This keeps the code better
readable as both properties can more easily distinguished.

Related file names have been adapted too.

This has become an extra commit instead of applying a fixup. This is due to the
fact that it is not simply possible to do this as the naming spreads over
multiple commits with moving parts. To not overwrite all commits one-by-one and
maintain reviewability, the compromise is a separate commit.