Enhance supported chain handling
As a part of keeping aligned to architectural decision while evolving
it for a better code reusing and in the same time keeping stick to SOC
design principle, we agree somehow that presentation layer (UI
components) are the one responsible for decising what elements should
be shown and what not, while composable is hiding business logic inside
and just responsible of sending values or packed methods for the UI to
use!
For this sake, variable names like `show` are no longer allowed because
it gives the impression of a UI action that is the components
responsability! so, if there's a variable like that, it will be
called `available`
On the other hand, according to the old decision of isolating
composables from each other and passing them as params from UI, and
the works for helper functions, a getter for supported chain has been
added to configuration store for this check!
For the sake of keeping commits as small as possible, I've just
applied the mentioned rule on supported chain and will continue with
this approach the next commit!