frontend: recover transfer composable
While doing the refactoring, it was intended to get rid of the `useTransfer`
composable as doing transfers is much easier now (from a Vue components
perspective). Unfortunately this is not the case. So a composable still makes
sense.
Unfortunately it was very hard to maintain the composable with fixup commits in
the history. For simplicity it got thereby removed and now recovered (and
adapted) in two commits.