frontend: Improve types of the properties of Transfer
This adds a sourceAmount and targetAmount (type: TokenAmount) to Transfer and removes the sourceToken, the targetToken and the amount.
The tokens can now be accessed from the amounts. This also enables later having a different target amount from the source one.
Also it makes the fees a TokenAmount because the fee model changes and we do not want to break backwards compatibility with encoding in the local storage.
For convenience an EthereumAmount class is created which inherits from the TokenAmount and sets the right symbol when created.
This type is also now used in the useRequestFee composable. Later, the composable will need to be adapted to the new fee model and use a TokenAmount.