frontend: add new transfer history store
The store is pretty simple so far. The most advanced part is the persistence to
the browsers local storage. Therefore a new plugin for Pinia has been added.
This specific one has been chosen as it allows to specifically define
a (de)serializer. This is necessary for us to properly encode and decode our
transfers. In fact this is finally the part where the whole work for the
encoding comes into effect.