frontend: add new lazy wrapper component
The LazyWrapper component is a quite special one as it makes use of the
intersection observer feature of modern browsers. Though it also supports
outdated browsers. It can be used to wrap a component so that it will not be
rendered before it becomes visible, or becomes close to be visible. This is
especially useful for long lists. This is even more important for components
that load data in background or similar.