frontend: add more advanced token generation
Instead of just generating the transfer data to instantiate a transfer, this new
test function is used to get transfers directly with certain higher level
properties that are hard to express by hand in the transfer data. It makes it
easy and nice to read that a completed, failed, ... transfer is needed.