frontend: adapt visual components to view transfer
This adapts some visual components that display transfers according to the new
design and the new data.

This makes use of the new Expandable component for the StatusTransfer. It shows
only a bare minimum of information while unexpended. Therefore it adds a colored
circle to visualize the current status of the transfer. Moreover it hides the
transfer progress steps as soon as a transfer got completed.

Also the shown transfer summary information got updated according to the new
design. Including a link to the block explorer for the request transaction.