frontend: add smart transfer grouping composable
According to the design, it is required to group transfers in the history into
different time windows. Like: today, last 3 days, last week, ... This new
composable is doing this. In can be configured with such time windows
individually and is fully reactive in all directions.
Though this is not yet used in the UI.