frontend: integrate transfer history component
This adds the `TransferHistory` component into the activity tab of the
`Home` view. Therefore it needs to make a few improvements on some
styling here and there.

It also removes the whole transfer status view from the `RequestDialog`.
Thereby it can rid of a lot of stuff and becomes quite slim and focused
on a single task. Therefore it got merge with the `useTransfer`
composable, as this composable has no more value as there are no
transfer errors etc.