frontend: allow to programmatically change tabs
This is an extension of the `Tabs` component that allows to switch the
tabs not only by the user with clicking the headers, but also
programmatically via a Vue component property.
Note that it has been actively decided to not use a `v-model` approach.
This makes the property optional. Nevertheless the component emits an
event to allow the parent to keep in sync which tab is currently active.