frontend: switch to activities on new transfer
When the user initiates a new transfer he will now automatically switch
to the activities tab where can see the new in progress transfer.

This new feature has been implemented via routing. The reasoning behind
this is better decoupling. In fact the `RequestDialog` component should
not not have any knowledge how to show the activities tab to the user.
It should not need to adapt or anything else if the UI composition
changes. I should also not know that it is embedded in a tabs component.
The only thing it should know/do: if a new transfer got created, call
a method to forward the user to the activities so he can follow the
progress. That's it.
The most obvious shared resource that can be used for this is the router
of the application. Moreover this specific behavior (show user the
activities) is put into a utility function. The purpose of this is that
we can alternate the exact routing (or even exchange it completely)
without affecting any component using this function.
Finally the component which is aware of the tabs and can control them is
the `Home` view. Therefore the current mechanism uses a query parameter
that gets observed by the view component. To keep functional in both
ways (switch tabs via user click as well as routing), the component must
keep the query parameter in sync with the emitted events of the `Tabs`
component. Else the user would do a transfer, get to the activity tab,
click on the transfer tabs and then could no more get automatically
presented with the activity tab again.