frontend: fix expansion of active transfer
When the user makes a new transfer and switches automatically to the
activities tab, the just created transfer should be expanded so the user
can follow the progress.
This features was already implemented earlier, but had one downside.
Once this transfer completed, it collapsed immediately. To support
a better user experience it should stay open, so the user can see the
summary. The progress steps will still disappear as for all completed
transfers.

The problem with this so far was that the feature got implemented in the
`TransferHistory` component via a property to the `TransferStatus`
component. If it is the first overall transfer in the list and is
active, it opens. In turn, once it is no more active, it automatically
closes.
As we have no unique identifiers for transfers before the request
transaction, it is hard to establish any better mechanism in the history
component to solve the problem. Therefore it got moved into the transfer
status component. If the transfer becomes active, it opens immediately
(in fact the user won't see at as he is not yet at the activity tab).
When the transfer is done, it will collapse after a 7s delay. The
closing is somehow necessary because else we end up with ugly cases
where the user does another transfer, gets back to the activity tab
again and two transfers are open now.