frontend: add wallet connection menu
This adds a new component that lists all possible wallet connection options to
the users. This replaces the old "button solution" and implements the new
design. This also finally moves the connectivity logic outside of the home view.

There is an issue currently with support for blur filters by all web-browsers.
See the code comments to get the details of the workaround.

Please note that there are currently no tests for this new component. This is
because we have no strategy for mocking/using composables yet. This is an open
issue that we will tackle soon. To get this feature ready, we do skip this as we
did before.