frontend: squash request signer composable
This composable is basically only used by the `useWallet` composable.
But it is a lot of effort to get this one composable into the other one.
The Vue components get hard to read with so much boilerplate.
Furthermore it makes the whole logic of the composable more complicated
which is simply not necessary.