frontend: improve wallet composable
The `useWallet` composable provides comfort to (re)connect a wallet in
combination of using the configuration, settings and provider store. One of the
functions it exposes is meant to reconnect to the wallet from the last session.
In fact the caller should just state: please reconnect to the last wallet if
there was one. And that's it. The same as when doing: please connect via
MetaMask. The composable gets access to all resources to manage the whole
process. There is no need to make an exception and have the caller set the
wallet after reconnection. This change simplifies the caller and the composable too.

This includes the fix for a typo of the `MockedMet[a]MaskProvider` class.