frontend: add new safe teleport vue component
We started to use the new Vue3 `Teleport` component. Unfortunately this
doesn't always work so well when having the portal within the Vue
component tree. The teleport fails because the portal is not available
yet. This for example happens when we add a button to the transfer
history that should be ported into the action button area. If the user
loads the activity tab first, the button does not show up and the
console shows an error. This does not happen for the request dialog, as
here we render the button conditionally on the signer availability.

To solve this problem properly, a new `SafeTeleport` component got
added. This component ensures that the portal target is actually
available in the DOM. Therefore it makes use of the MutationObserver of
modern web-browsers. In case that is not available, it falls back to
a simple periodic check.
Note that it is not possible to use the `disabled` property of the
`Teleport` component here as this would still lead to a failure.